AspectRatio moved to API package
