+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/ctrl/ImageAttributesDTOTest.java
@@ -10,6 +10,8 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import de.lexasoft.mandelbrot.api.AspectRatio;
+
 /**
  * @author admin
  *
@@ -34,7 +36,7 @@ class ImageAttributesDTOTest {
 		assertEquals(459, cut.getImageWidth());
 		assertEquals(405, cut.getImageHeight());
 		assertNull(cut.getImageFilename());
-		assertEquals(AspectRatioDTO.FILL, cut.getAspectRatioHandle());
+		assertEquals(AspectRatio.FILL, cut.getAspectRatioHandle());
 	}
 
 }
