+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/ctrl/MandelbrotAttributesDTOTest.java
@@ -18,6 +18,7 @@ import org.junit.jupiter.api.Test;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 
+import de.lexasoft.mandelbrot.api.AspectRatio;
 import de.lexasoft.mandelbrot.api.ColorGradingStyle;
 import de.lexasoft.mandelbrot.api.PaletteVariant;
 import de.lexasoft.mandelbrot.api.TransitionVariant;
@@ -56,7 +57,7 @@ class MandelbrotAttributesDTOTest {
 		assertEquals(4590, cut.getImage().getImageWidth());
 		assertEquals(4050, cut.getImage().getImageHeight());
 		assertEquals("./junit-tmp/mandelbrot-test.tiff", cut.getImage().getImageFilename());
-		assertEquals(AspectRatioDTO.FILL, cut.getImage().getAspectRatioHandle());
+		assertEquals(AspectRatio.FILL, cut.getImage().getAspectRatioHandle());
 
 		// Color
 		assertSame(PaletteVariant.CUSTOM, cut.getColor().getPaletteVariant());
